Clean up / move back some contants
